# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RideAuto API (публичные эндпоинты)
  version: "1.0"
  description: |
    Минимальное описание read-only API каталога и служебных маршрутов.
    Полный набор query-параметров для `/api/cars` и ответы см. в `backend/fastapi_app/routers/search.py` и схемах Pydantic.
    Лимиты GET по IP: `WRA_RATE_LIMIT_GET_CARS_PER_MINUTE`, `WRA_RATE_LIMIT_GET_FACETS_PER_MINUTE` (0 = без лимита).
    Курсорная пагинация: query `cursor` в `/api/cars` для сортировок date_*, year_*, mileage_*, price_high / price_low; см. `meta.next_cursor`.
    Без `cursor`: глубокий `page` ограничен для анонимов (`WRA_CATALOG_MAX_PAGE_OFFSET_ANON`, по умолчанию 80; `-1` = выкл.); при сессии Bearer — `WRA_CATALOG_MAX_PAGE_OFFSET_AUTH` (0 = без лимита).
    Кэш фасетов в воркере: `WRA_FACETS_CACHE_TTL` (сек, по умолчанию 300).
    Метрики: `GET /metrics` (legacy alias: `GET /api/metrics`) при `WRA_METRICS_ENABLED=1`
    (legacy env alias: `WRA_PROMETHEUS_METRICS=1`).
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    cookieAuth:
      type: apiKey
      in: cookie
      name: wra_session
      description: Сессия magic-link (WRA_AUTH_COOKIE_NAME, по умолчанию wra_session)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: unauthorized
...
paths:
...
  /api/push/vapid-public-key:
    get:
      summary: Публичный VAPID-ключ для Web Push
      description: Требует `WRA_PUSH_VAPID_PUBLIC_KEY` на сервере; иначе 503 `push_not_configured`.
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [public_key]
                properties:
                  public_key: { type: string }
        "503":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/push/subscribe:
    post:
      summary: Сохранить Web Push подписку (авторизованный пользователь)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, keys]
              properties:
                endpoint: { type: string, minLength: 20, maxLength: 2000 }
                keys:
                  type: object
                  required: [p256dh, auth]
                  properties:
                    p256dh: { type: string }
                    auth: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: Удалить Web Push подписку по endpoint
      security:
        - cookieAuth: []
      parameters:
        - name: endpoint
          in: query
          required: true
          schema: { type: string, minLength: 20 }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/subscriptions:
    get:
      summary: Подписки
      security:
        - bearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema: { type: object }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: Создать подписку
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                filters: { type: object }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/subscriptions/{sid}:
    patch:
      summary: Обновить подписку (notify / name)
      security:
        - cookieAuth: []
      parameters:
        - name: sid
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                notify_enabled: { type: boolean }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: Удалить подписку
      security:
        - bearerAuth: []
      parameters:
        - name: sid
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/subscriptions/run-notifications:
    post:
      summary: Рассылка по подпискам (X-Admin-Key)
      parameters:
        - name: X-Admin-Key
          in: header
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { type: object }
        "403":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }